+++ custom_components/linux_audio_server/__init__.py
@@ -55,6 +55,7 @@ PLATFORMS: list[Platform] = [
     Platform.SELECT,
     Platform.BUTTON,
     Platform.DEVICE_TRACKER,
+    Platform.NUMBER,
 ]